Move argument checking library code to NodeValueOps
